+++ components/raw/embl-notifications/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.0-alpha.2
 
 * Improved formatting support and conditional logic https://github.com/visual-framework/vf-core/pull/817
+* Notification cookies and memory options
 
 ## 1.0.0-alpha.1 (2020-03-20)